Se restructuró el nuevo cotizador para una mejor organización y se toma como herencia el existente
