+++ public/plantilla/plataforma2.html
@@ -587,8 +587,8 @@
                   </div>
   
                   <div class="d-flex justify-content-end">
-                    <button id="boton_cotizar_2"  onclick="cotizador()" 
-                    class="btn btn-primary btn-sm mx-2">Cotizar envío</button>
+                    <button id="boton_cotizar_2"
+                    class="btn btn-primary btn-sm mx-2 cotizador-button cotizador-2">Cotizar envío</button>
                     
                   </div>
 
@@ -676,8 +676,8 @@
               </div> -->
 
               </form>
-              <button id="boton_cotizar"  onclick="cotizador()" 
-              class="btn btn-primary btn-block btn-lg mb-3">
+              <button id="boton_cotizar"
+              class="btn btn-primary btn-block btn-lg mb-3 cotizador-button">
                 Cotizar envío
               </button>
